add EigenAddressCliError type and reorganize egnaddrs crate (remove egnaddr bin)
