+++ crates/eigen-cli/src/main.rs
@@ -1,10 +1,10 @@
 use clap::Parser;
-use eigen_cli::{args::Args, get_addresses};
+use eigen_cli::{args::Args, eigen_address::EigenAddresses};
 
 #[tokio::main]
 async fn main() {
     let args = Args::parse();
-    let addresses = get_addresses(args).await.unwrap();
+    let addresses = EigenAddresses::get_addresses(args).await.unwrap();
 
     println!("{}", serde_json::to_string_pretty(&addresses).unwrap());
 }
